pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inference-agent"
version = "0.1.0"
description = "LLM inference benchmark agent — automated vLLM/SGLang configuration optimizer"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2",
    "openai>=1.50",
    "docker>=7.0",
    "aiohttp>=3.9",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "prometheus-client>=0.20",
    "huggingface-hub>=0.20",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.35",
    "plotly>=5.0",
    "pandas>=2.0",
    "psycopg[binary]>=3.1",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "testcontainers[postgres]>=4.0",
]

[project.scripts]
inference-agent = "inference_agent.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: integration tests requiring Docker (Postgres testcontainer)",
]

[tool.hatch.build.targets.wheel]
packages = ["src/inference_agent"]
